#40 | Remove username from package naming in project
