+++ build.gradle.kts
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id("com.google.dagger.hilt.android") version "2.48" apply false
-    id("com.rockar.android.plugins.verify")
-    id("com.rockar.android.plugins.format")
-    id("com.rockar.android.plugins.coverage")
+    id("com.android.plugins.verify")
+    id("com.android.plugins.format")
+    id("com.android.plugins.coverage")
 }
