+++ buildSrc/src/main/kotlin/com/android/marvelapp/utils/AndroidGradleExtensions.kt
@@ -1,0 +1,28 @@
+package com.android.marvelapp.utils
+
+import com.android.build.gradle.LibraryExtension
+import org.gradle.api.Action
+import org.gradle.api.Project
+import org.gradle.api.artifacts.Dependency
+import org.gradle.api.artifacts.dsl.DependencyHandler
+import org.gradle.api.plugins.ExtensionAware
+import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
+
+val Project.androidLibrary: LibraryExtension?
+    get() =
+        (this as ExtensionAware).extensions.findByName("android") as? LibraryExtension
+
+fun Project.android(configure: Action<LibraryExtension>): Unit =
+    (this as ExtensionAware).extensions.configure("android", configure)
+
+fun LibraryExtension.kotlinOptions(configure: Action<KotlinJvmOptions>): Unit =
+    (this as ExtensionAware).extensions.configure("kotlinOptions", configure)
+
+fun DependencyHandler.implementation(dependencyNotation: Any): Dependency? =
+    add("implementation", dependencyNotation)
+
+fun DependencyHandler.kapt(dependencyNotation: Any): Dependency? =
+    add("kapt", dependencyNotation)
+
+fun DependencyHandler.lintChecks(dependencyNotation: Any): Dependency? =
+    add("lintChecks", dependencyNotation)
