+++ buildSrc/src/main/kotlin/com/android/marvelapp/utils/StringUtils.kt
@@ -1,0 +1,11 @@
+package com.android.marvelapp.utils
+
+fun String.capitalizeFirst(): String {
+    return this.replaceFirstChar {
+        if (it.isLowerCase()) {
+            it.titlecase()
+        } else {
+            it.toString()
+        }
+    }
+}
