--- buildSrc/src/main/kotlin/com/rockar/android/marvelapp/dependencies/DependencyVersions.kt
@@ -1,23 +1,0 @@
-package com.rockar.android.marvelapp.dependencies
-
-const val hiltAndroidPluginVersion = "2.48"
-
-const val coreKtxVersion = "1.9.0"
-const val appcompatVersion = "1.6.1"
-const val materialVersion = "1.8.0"
-const val coroutinesAndroidVersion = "1.3.9"
-
-const val retrofitVersion = "2.9.0"
-const val okHttpBomVersion = "4.11.0"
-
-const val jUnitVersion = "4.13.2"
-const val jUnitInstrumentedVersion = "1.1.5"
-const val espressoCoreVersion = "3.5.1"
-const val truthVersion = "1.1.4"
-const val mockkVersion = "1.13.7"
-
-object GradleConfigVersions {
-    const val compileSdk = 34
-    const val minSdk = 28
-    const val jvmTarget = "17"
-}
