--- buildSrc/src/main/kotlin/com/rockar/android/marvelapp/utils/StringUtils.kt
@@ -1,11 +1,0 @@
-package com.rockar.android.marvelapp.utils
-
-fun String.capitalizeFirst(): String {
-    return this.replaceFirstChar {
-        if (it.isLowerCase()) {
-            it.titlecase()
-        } else {
-            it.toString()
-        }
-    }
-}
