+++ features/core/rules/src/main/java/com/android/rules/detector/registry/CustomIssueRegistry.kt
@@ -1,0 +1,19 @@
+package com.android.rules.detector.registry
+
+import com.android.rules.detector.ExcludedClassIfHiltModuleRule
+import com.android.tools.lint.client.api.IssueRegistry
+import com.android.tools.lint.client.api.Vendor
+import com.android.tools.lint.detector.api.CURRENT_API
+import com.android.tools.lint.detector.api.Issue
+
+class CustomIssueRegistry : IssueRegistry() {
+
+    override val api: Int
+        get() = CURRENT_API
+
+    override val issues: List<Issue>
+        get() = listOf(ExcludedClassIfHiltModuleRule.ISSUE)
+
+    override val vendor: Vendor
+        get() = Vendor("HeroesApp")
+}
